Add MongoDB support and refactor repository classes
- Introduced a new project for MongoDB repositories in the solution.
- Updated `IEntityRepository` and `QueryResult` to use `long` for count properties.
- Refactored `AuditableEntityRepository` and `EntityRepository` to use generics more effectively.
- Enhanced error logging in repository methods.
- Implemented `MongoDBContext` and `MongoDBOptions` for configuration management.
- Added service extension methods for MongoDB repository registration.
- Updated `Program.cs` and `appsettings.json` for MongoDB integration.
- Introduced global usings for improved code readability.
- Cleaned up project references and updated package dependencies.

+++ src/FluentCMS.Core.Repositories.Abstractions/IEntityRepository.cs
@@ -4,7 +4,7 @@ public interface IEntityRepository<T> where T : class, IEntity
 {
     Task<IEnumerable<T>> GetAll(CancellationToken cancellationToken = default);
     Task<QueryResult<T>> Query(QueryOptions<T> options, CancellationToken cancellationToken = default);
-    Task<int> Count(Expression<Func<T, bool>>? filter = default, CancellationToken cancellationToken = default);
+    Task<long> Count(Expression<Func<T, bool>>? filter = default, CancellationToken cancellationToken = default);
     Task<T> Add(T entity, CancellationToken cancellationToken = default);
     Task<T> Update(T entity, CancellationToken cancellationToken = default);
     Task<T> Remove(T entity, CancellationToken cancellationToken = default);
